+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 to rendering artifacts)
 - Fix rendering whitespace characters causing repeated font creation
 - Fix underline rendering cutting off sometimes (made it so underline always renders within cell)
+- Fix keyboard input: F13-F24 and other extended keys now work, emoji input via picker supported
 
 ## [0.2.2] - 2025-05-08